moved header logic into its own component
