+++ acp-webapp/app/auth/sign-up/page.tsx
@@ -1,4 +1,4 @@
-import { SignUpForm } from '@/components/sign-up-form'
+import { SignUpForm } from '@/components/login/sign-up-form'
 
 export default function Page() {
   return (
