+++ acp-webapp/app/auth/update-password/page.tsx
@@ -1,4 +1,4 @@
-import { UpdatePasswordForm } from '@/components/update-password-form'
+import { UpdatePasswordForm } from '@/components/login/update-password-form'
 
 export default function Page() {
   return (
